requirements.txt
google-cloud-aiplatform==1.69.0
python-dotenv==1.0.1
pytest==7.4.3
pytest-cov==2.11.1
streamlit==1.27.2
google-cloud-aiplatform
google-cloud-bigquery
python-dotenv
